data-platform/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
httpx==0.27.2
redis[asyncio]==5.0.8
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
apscheduler==3.10.4
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
